:bread: :art: Reformat components preparing Pull Request
- Rename files like login_mobile.dart to login.mobile.dart
- Deleted main.dart
- Deleted images
- Reformat entire code

# Future
- Needs to refactor all components to be generic
    - Some widgets are reimplemented, refactor it
- Change comments to english
