+++ lib/src/mobile/drawer/drawer_item.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 class DrawerItem extends StatelessWidget {
   final String title;
   final IconData icon;
+
   const DrawerItem(this.title, this.icon);
 
   @override
