+++ pubspec.yaml
@@ -17,7 +17,7 @@ dependencies:
     sdk: flutter
   crypto: ^2.1.5
   device_info: ^1.0.0
-  responsive_builder: ^0.1.2
+  responsive_builder: ^0.3.0
 
 dev_dependencies:
   flutter_test:
@@ -27,7 +27,6 @@ dev_dependencies:
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 # The following section is specific to Flutter.
-
 
 flutter:
   # To add custom fonts to your application, add a fonts section here,
